pyproject.toml
[project]
name = "pyprograms"
version = "0.1.0"
description = "Sistema Multi-Agente ADK para Control de Riego IoT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.5.0",
    "paho-mqtt>=2.0.0",
    "python-dotenv>=1.0.0",
    "mcp",
    "httpx",
    "nest_asyncio",
]
